refactor: DailyMovieCollectionView, DailyBoxOfficeCollectionView 타입 삭제, 해당 타입에서 수행하던 기능 뷰컨트롤러로 이전
