+++ BoxOffice/@iOS 14.0 ListCell/DailyMovieViewController.swift
@@ -11,15 +11,22 @@ import UIKit
 final class DailyMovieViewController: UIViewController {
     private var networkManager = NetworkManager()
     private var boxOfficeEndPoint: BoxOfficeEndPoint?
-    
     private var refreshControl = UIRefreshControl()
-    
-    lazy private var collectionView = DailyMovieCollectionView(frame: view.bounds, collectionViewLayout: createLayout())
+    private var dailyBoxOffice: DailyBoxOffice?
+    private typealias DataSource = UICollectionViewDiffableDataSource<Section, DailyBoxOffice.BoxOfficeResult.Movie>
+    private var movieDataSource: DataSource?
+    lazy private var collectionView: UICollectionView = {
+        let collectionView = UICollectionView(frame: view.bounds, collectionViewLayout: createLayout())
+        collectionView.autoresizingMask = [.flexibleWidth, .flexibleHeight]
+        collectionView.register(DailyBoxOfficeCollectionViewCell.self, forCellWithReuseIdentifier: DailyBoxOfficeCollectionViewCell.reuseIdentifier)
+        view.addSubview(collectionView)
+        
+        return collectionView
+    }()
     
     override func viewDidLoad() {
         super.viewDidLoad()
         
-        view.addSubview(collectionView)
         collectionView.delegate = self
         refreshData()
         refreshControl.addTarget(self, action: #selector(refreshData), for: .valueChanged)
@@ -50,20 +57,49 @@ final class DailyMovieViewController: UIViewController {
     
     private func fetchDailyBoxOfficeData() {
         guard let endPoint = boxOfficeEndPoint else { return }
+        
         networkManager.request(endPoint: endPoint, returnType: DailyBoxOffice.self) { [weak self] in
             switch $0 {
             case .failure(let error):
                 print(error)
             case .success(let result):
-                self?.collectionView.dailyBoxOffice = result
+                self?.dailyBoxOffice = result
                 
                 DispatchQueue.main.async {
-                    self?.collectionView.setupDataSource()
-                    self?.collectionView.setupSnapshot()
+                    self?.setupDataSource()
+                    self?.setupSnapshot()
                     self?.refreshControl.endRefreshing()
                 }
             }
         }
+    }
+}
+
+fileprivate enum Section: Hashable {
+    case main
+}
+
+@available(iOS 14.0, *)
+extension DailyMovieViewController {
+    func setupDataSource() {
+        let cellRegistration = UICollectionView.CellRegistration<DailyMovieListCell, DailyBoxOffice.BoxOfficeResult.Movie> { (cell, indexPath, item) in
+            cell.updateWithItem(item)
+            cell.accessories = [.disclosureIndicator()]
+        }
+        
+        movieDataSource = DataSource(collectionView: collectionView) {
+            (collectionView: UICollectionView, indexPath: IndexPath, item: DailyBoxOffice.BoxOfficeResult.Movie) -> UICollectionViewCell? in
+            return collectionView.dequeueConfiguredReusableCell(using: cellRegistration, for: indexPath, item: item)
+        }
+    }
+    
+    func setupSnapshot() {
+        var snapshot = NSDiffableDataSourceSnapshot<Section, DailyBoxOffice.BoxOfficeResult.Movie>()
+        snapshot.appendSections([.main])
+        if let dailyBoxOffice = dailyBoxOffice {
+            snapshot.appendItems(dailyBoxOffice.boxOfficeResult.boxOfficeList, toSection: .main)
+        }
+        movieDataSource?.apply(snapshot, animatingDifferences: false)
     }
 }
 
@@ -79,18 +115,5 @@ extension DailyMovieViewController {
 extension DailyMovieViewController: UICollectionViewDelegate {
     func collectionView(_ collectionView: UICollectionView, didSelectItemAt indexPath: IndexPath) {
         collectionView.deselectItem(at: indexPath, animated: true)
-    }
-}
-
-@available(iOS 14.0, *)
-extension UIConfigurationStateCustomKey {
-    static let movieKey = UIConfigurationStateCustomKey("movie")
-}
-
-@available(iOS 14.0, *)
-extension UICellConfigurationState {
-    var movie: DailyBoxOffice.BoxOfficeResult.Movie? {
-        get { return self[.movieKey] as? DailyBoxOffice.BoxOfficeResult.Movie }
-        set { self[.movieKey] = newValue }
     }
 }
