+++ BoxOffice/Application/SceneDelegate.swift
@@ -2,7 +2,7 @@
 //  SceneDelegate.swift
 //  BoxOffice
 //
-//  Created by kjs on 13/01/23.
+//  Created by 리지, kokkilE on 2023/03/28.
 //
 
 import UIKit
@@ -14,7 +14,7 @@ class SceneDelegate: UIResponder, UIWindowSceneDelegate {
     func scene(_ scene: UIScene, willConnectTo session: UISceneSession, options connectionOptions: UIScene.ConnectionOptions) {
         guard let windowScene = (scene as? UIWindowScene) else { return }
         window = UIWindow(windowScene: windowScene)
-        let dailyBoxOfficeViewController = DailyMovieViewController()
+        let dailyBoxOfficeViewController = DailyBoxOfficeViewController()
         
         let navigationController = UINavigationController(rootViewController: dailyBoxOfficeViewController)
